unify plugins to Plugin.GHCSpecter
