+++ src/Plugin/GHCSpecter.hs
@@ -1,7 +1,7 @@
 {-# LANGUAGE BangPatterns #-}
 
 -- This module provides the current module under compilation.
-module Plugin.Timing
+module Plugin.GHCSpecter
   ( -- * main plugin entry point
 
     -- NOTE: The name "plugin" should be used as a GHC plugin.
@@ -24,12 +24,17 @@ import Control.Concurrent.STM
   )
 import Control.Monad (when)
 import Control.Monad.IO.Class (liftIO)
+import Data.Char (isAlpha)
 import Data.Foldable (for_)
+import Data.IORef (readIORef)
 import Data.IntMap (IntMap)
 import Data.IntMap qualified as IM
+import Data.List qualified as L
 import Data.Map (Map)
 import Data.Map qualified as M
 import Data.Maybe (mapMaybe)
+import Data.Set (Set)
+import Data.Set qualified as S
 import Data.Text qualified as T
 import Data.Time.Clock (getCurrentTime)
 import Data.Tuple (swap)
@@ -52,6 +57,20 @@ import GHC.Driver.Plugins
     defaultPlugin,
     type CommandLineOption,
   )
+import GHC.Driver.Session (DynFlags, getDynFlags)
+import GHC.Plugins
+  ( ModSummary,
+    Name,
+    localiseName,
+    showSDoc,
+  )
+import GHC.Tc.Types (TcGblEnv (..), TcM)
+import GHC.Types.Name.Reader
+  ( GlobalRdrElt (..),
+    GreName (..),
+    ImpDeclSpec (..),
+    ImportSpec (..),
+  )
 import GHC.Unit.Module.Graph
   ( ModuleGraph,
     ModuleGraphNode (..),
@@ -64,10 +83,11 @@ import GHC.Unit.Module.ModSummary
   )
 import GHC.Unit.Module.Name (moduleNameString)
 import GHC.Unit.Types (GenModule (moduleName))
+import GHC.Utils.Outputable (Outputable (ppr))
 import System.Directory (doesFileExist)
 import System.IO.Unsafe (unsafePerformIO)
 import Toolbox.Channel
-  ( ChanMessage (CMSession, CMTiming),
+  ( ChanMessage (CMCheckImports, CMSession, CMTiming),
     ChanMessageBox (..),
     ModuleGraphInfo (..),
     ModuleName,
@@ -81,12 +101,19 @@ import Toolbox.Util (showPpr)
 
 plugin :: Plugin
 plugin =
-  defaultPlugin {driverPlugin = driver}
+  defaultPlugin
+    { driverPlugin = driver
+    , typeCheckResultAction = typecheckPlugin
+    }
 
 -- shared across the session
 sessionRef :: TVar SessionInfo
 {-# NOINLINE sessionRef #-}
 sessionRef = unsafePerformIO (newTVarIO (SessionInfo Nothing emptyModuleGraphInfo))
+
+--
+-- driver plugin
+--
 
 getModuleNameText :: ModSummary -> T.Text
 getModuleNameText = T.pack . moduleNameString . moduleName . ms_mod
@@ -178,3 +205,72 @@ driver opts env = do
       hooks' = hooks {runPhaseHook = Just runPhaseHook'}
       env' = env {hsc_hooks = hooks'}
   pure env'
+
+--
+-- typechecker plugin
+--
+
+formatName :: DynFlags -> Name -> String
+formatName dflags name =
+  let str = showSDoc dflags . ppr . localiseName $ name
+   in case str of
+        (x : _) ->
+          -- NOTE: As we want to have resultant text directly copied and pasted to
+          --       the source code, the operator identifiers should be wrapped with
+          --       parentheses.
+          if isAlpha x
+            then str
+            else "(" ++ str ++ ")"
+        _ -> str
+
+formatImportedNames :: [String] -> String
+formatImportedNames names =
+  case fmap (++ ",\n") $ L.sort names of
+    l0 : ls ->
+      let l0' = "  ( " ++ l0
+          ls' = fmap ("    " ++) ls
+          footer = "  )"
+       in concat ([l0'] ++ ls' ++ [footer])
+    _ -> "  ()"
+
+mkModuleNameMap :: GlobalRdrElt -> [(ModuleName, Name)]
+mkModuleNameMap gre = do
+  spec <- gre_imp gre
+  case gre_name gre of
+    NormalGreName name -> do
+      let modName = T.pack . moduleNameString . is_mod . is_decl $ spec
+      pure (modName, name)
+    -- TODO: Handle the record field name case correctly.
+    FieldGreName _ -> []
+
+-- | First argument in -fplugin-opt is interpreted as the socket file path.
+--   If nothing, do not try to communicate with web frontend.
+typecheckPlugin ::
+  [CommandLineOption] ->
+  ModSummary ->
+  TcGblEnv ->
+  TcM TcGblEnv
+typecheckPlugin opts modsummary tc = do
+  dflags <- getDynFlags
+  usedGREs :: [GlobalRdrElt] <-
+    liftIO $ readIORef (tcg_used_gres tc)
+  let moduleImportMap :: Map ModuleName (Set Name)
+      moduleImportMap =
+        L.foldl' (\(!m) (modu, name) -> M.insertWith S.union modu (S.singleton name) m) M.empty $
+          concatMap mkModuleNameMap usedGREs
+
+  let rendered =
+        unlines $ do
+          (modu, names) <- M.toList moduleImportMap
+          let imported = fmap (formatName dflags) $ S.toList names
+          [T.unpack modu, formatImportedNames imported]
+
+  let modName = T.pack $ moduleNameString $ moduleName $ ms_mod modsummary
+  case opts of
+    ipcfile : _ -> liftIO $ do
+      socketExists <- doesFileExist ipcfile
+      when socketExists $
+        runClient ipcfile $ \sock ->
+          sendObject sock (CMBox (CMCheckImports modName (T.pack rendered)))
+    _ -> pure ()
+  pure tc
